requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
chromadb==0.6.3
sentence-transformers==3.4.1
requests==2.32.3
python-dotenv==1.0.1
groq==0.14.0
langchain-core==0.3.35
langgraph==0.2.74
